+++ printcetnrum_web.client/src/app/pages/diploma-page/diploma-page.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 import { SnackBarUtil } from '../../shared/snackbar-util';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { DiplomaTextBox } from '../../models/diploma-container.model';
-import {DesignFilesHandlerService} from "../../services/design-files-handler.service";
+import {DesignFilesHandlerService} from "../../services/file-services/design-files-handler.service";
 
 @Component({
   selector: 'app-diploma-page',
