+++ printcetnrum_web.client/src/app/services/auth-services/reset-password.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { ResetPassword } from '../models/reset-password.model';
+import { ResetPassword } from '../../models/reset-password.model';
 
 @Injectable({
   providedIn: 'root'
