+++ printcetnrum_web.client/src/app/services/file-services/design-files-handler.service.ts
@@ -1,8 +1,8 @@
 import {inject, Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {AuthService} from "./auth.service";
+import {AuthService} from "../auth-services/auth.service";
 import {Observable} from "rxjs";
-import {UserFile} from "../models/user-file";
+import {UserFile} from "../../models/user-file";
 
 @Injectable({
   providedIn: 'root'
