+++ printcetnrum_web.client/src/app/services/order-services/order.service.ts
@@ -1,8 +1,8 @@
 import { inject, Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Order, OrderItem } from '../models/order-models/order.model';
-import { AuthService } from './auth.service';
+import { Order, OrderItem } from '../../models/order-models/order.model';
+import { AuthService } from '../auth-services/auth.service';
 
 @Injectable({
   providedIn: 'root',
